geodashml/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
configparser==5.0.0
docker-pycreds==0.4.0
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.28.1
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.2
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
pathtools==0.1.2
Pillow==7.1.2
promise==2.3
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
wandb==0.8.35
watchdog==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
